Cars_root/requirements.txt
asgiref==3.5.1
attrs==21.4.0
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-filter==21.1
django-rest-swagger==2.2.0
django-soft-delete==0.9.21
django-url-filter==0.3.15
djangorestframework==3.13.1
drf-yasg==1.20.0
enum-compat==0.0.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
